+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.6-preview] - 2019-10-23
+
+### Added
+* Added `CanBuild` property to all build targets.
+
+### Changed
+* Updated `com.unity.platforms` to `0.1.6-preview`.
+
 ## [0.1.5-preview] - 2019-10-22
 
 ### Changed